[feat] Separate DepthDropdown and DepthSlider components
