+++ aim/web/ui/src/components/DepthDropdown/DepthDropdown.d.ts
@@ -1,0 +1,6 @@
+export interface IDepthDropdownProps {
+  index: number;
+  pathValue: string | string[];
+  depth: number;
+  onDepthChange: (value: number, index: number) => void;
+}
