+++ aim/web/ui/src/components/DepthDropdown/DepthDropdown.scss
@@ -1,0 +1,36 @@
+@use 'src/styles/abstracts' as *;
+
+.DepthDropdown {
+  &__button {
+    display: inline-flex;
+    align-items: center;
+    margin-left: $space-xxxs;
+  }
+  &__autocomplete {
+    &__popper{
+      width: auto!important;
+    }
+    &__select {
+      width: 100%;
+      & > input {
+        padding: $space-xs;
+        font-size: $text-sm;
+        border-bottom: $border-main;
+      }
+      &__optionIcon {
+        margin-left: $space-xxxs;
+        color: $primary-color;
+      }
+      &__optionLabel {
+        margin-right: auto;
+        overflow: hidden;
+        text-overflow: ellipsis;
+        white-space: nowrap;
+        &.selected {
+          color: $primary-color;
+        }
+      }
+    }
+  }
+}
+
