+++ aim/web/ui/src/components/DepthDropdown/index.tsx
@@ -1,0 +1,5 @@
+import DepthDropdown from './DepthDropdown';
+
+export * from './DepthDropdown.d';
+
+export default DepthDropdown;
