+++ aim/web/ui/src/components/DepthSlider/DepthSlider.d.ts
@@ -1,0 +1,6 @@
+export interface IDepthSliderProps {
+  index: number;
+  pathValue: string | string[];
+  depth: number;
+  onDepthChange: (value: number, index: number) => void;
+}
