+++ aim/web/ui/src/components/DepthSlider/index.tsx
@@ -1,0 +1,5 @@
+import DepthSlider from './DepthSlider';
+
+export * from './DepthSlider.d';
+
+export default DepthSlider;
