+++ aim/web/ui/src/components/MediaSet/MediaSet.scss
@@ -27,39 +27,6 @@
     }
     &__path {
       display: inline-flex;
-      &__depthDropdown{
-        &__button {
-          display: inline-flex;
-          align-items: center;
-          margin-left: $space-xxxs;
-        }
-        &__autocomplete {
-          &__popper{
-            width: auto!important;
-          }
-          &__select {
-            width: 100%;
-            & > input {
-              padding: $space-xs;
-              font-size: $text-sm;
-              border-bottom: $border-main;
-            }
-            &__optionIcon {
-              margin-left: $space-xxxs;
-              color: $primary-color;
-            }
-            &__optionLabel {
-              margin-right: auto;
-              overflow: hidden;
-              text-overflow: ellipsis;
-              white-space: nowrap;
-              &.selected {
-                color: $primary-color;
-              }
-            }
-          }
-        }
-      }
       &__title {
         padding-bottom: 0.125rem;
         font-size: $text-sm;
@@ -70,7 +37,7 @@
           cursor: pointer;
         }
         &__key, &__value {
-          &.slider {
+          &.stacked {
             overflow: hidden;
             text-overflow: ellipsis;
             white-space: nowrap;
@@ -79,7 +46,7 @@
         }
         &__key {
           margin-right: $space-xxxs;
-          &.slider {
+          &.stacked {
             max-width: toRem(137px);
           }
         }
@@ -88,30 +55,9 @@
           overflow: hidden;
           text-overflow: ellipsis;
           white-space: nowrap;
-          &.slider {
+          &.stacked {
             max-width: toRem(227px);
           }
-        }
-      }
-    }
-    &__sliderContainer {
-      display: flex;
-      align-items: center;
-      padding-top: $space-xxs;
-      & > .Slider {
-        width: toRem(407px);
-        height: 100%;
-        & .MuiSlider-mark {
-          top: 50%;
-          transform: translate(-1px, -50%);
-          width: toRem(1px);
-          height: toRem(5px);
-        }
-        & .prevIconBtn {
-          margin-right: $space-xxs;
-        }
-        & .nextIconBtn {
-          margin-left: $space-xxs;
         }
       }
     }
